Se agrega autorizacion de roles, productos, usuarios y sus respectivas rutas

+++ Documents/JERFFRE - Personal/ORT/Cuatrimestre 4 - 2024/TP2 - Taller de Programación 2/TP2-TPFINAL/controllers/ProductController.js
@@ -4,10 +4,26 @@ class ProductController {
 
   productService = new ProductService();
   
+  createProductService = async (req, res, next) => {
+    try {
+      const { name, precio, stock } = req.body;
+      const newProduct = await this.productService.createUserService({
+        name,
+        precio,
+        stock,
+      });
+      res.status(200).send({ success: true, message: newProduct });
+    } catch (error) {
+      res.status(400).send({ success: false, message: error.message });
+    }
+  };
+
+
+
   getAllProductService = async (req, res) => {
     try {
-      const data = await this.productService.getAllProductService();
-      res.status(200).send({ success: true, message: data });
+      const products = await this.productService.getAllProductService({ include: { model: User, as: 'creator' } });
+      res.status(200).send({ success: true, message: products });
     } catch (error) {
       res.status(400).send({ success: false, message: error.message });
     }
@@ -16,43 +32,45 @@ class ProductController {
   getProductByIdService = async (req, res) => {
     try {
       const { id } = req.params;
-      const data = await this.productService.getProductByIdService(id);
-      res.status(200).send({ success: true, message: data });
+      const product = await this.productService.getProductByIdService(id);
+      res.status(200).send({ success: true, message: product });
     } catch (error) {
       res.status(400).send({ success: false, message: error.message });
     }
   };
-  createProductService = async (req, res, next) => {
-    try {
-      const { id, name, precio, stock } = req.body;
-      const data = await this.productService.createUserService({
-        id,
-        name,
-        precio,
-        stock,
-      });
-      res.status(200).send({ success: true, message: data });
-    } catch (error) {
-      res.status(400).send({ success: false, message: error.message });
-    }
-  };
+ 
   
   updateProductService = async (req, res) => {
     try {
       const { name, precio, stock} = req.body;
-      console.log(`🚀 ~ UserController ~ updateUser= ~ pass:`, pass)
       const { id } = req.params;
-      const data = await this.productService.updateUserService({id, name, precio, stock});
-      res.status(200).send({ success: true, message: data });
+
+      const previousProduct = await Producto.findByPk(id);
+      if (!previousProduct) {
+        return res.status(404).json({ message: "Producto no encontrado." });
+      }
+
+      const updatedProduct = await this.productService.updateUserService({id, name, precio, stock});
+      res.status(200).json(updatedProduct);
+
     } catch (error) {
       res.status(400).send({ success: false, message: error.message });
     }
   };
+
+
+
   deleteProductService = async (req, res) => {
     try {
       const { id } = req.params;
-      const data = await this.productService.deleteProductService(id);
-      res.status(200).send({ success: true, message: data });
+
+      const searchedProduct = await Producto.findByPk(id);
+      if (!searchedProduct) {
+        return res.status(404).json({ message: "Producto no encontrado." });
+      }
+
+      const product = await this.productService.deleteProductService(id);
+      res.status(200).send({ success: true, message: product });
     } catch (error) {
       res.status(400).send({ success: false, message: error.message });
     }
